requirements.txt
arrow==1.2.3
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
defusedxml==0.7.1
Django==4.1.3
django-extensions==3.2.1
django-geojson==4.0.0
django-gisserver==1.2.7
django-leaflet==0.29.0
Fiona==1.8.22
Flask==2.2.2
GDAL @ file:///C:/Users/cfick/dev/dierenhulp/GDAL-3.4.3-cp310-cp310-win_amd64.whl
geographiclib==2.0
geopy==2.3.0
geos==0.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
lru-dict==1.2.0
lxml==4.9.1
MarkupSafe==2.1.1
munch==2.5.0
orjson==3.9.1
Pillow==9.3.0
proj==0.2.0
pycountry==22.3.5
python-dateutil==2.8.2
python-dotenv==0.21.0
six==1.16.0
spatialite==0.0.3
sqlparse==0.4.3
tzdata==2022.6
Werkzeug==2.2.2
